+++ i_sound.h
@@ -230,7 +230,7 @@ dboolean I_MusicIsPlaying(void);
 extern int snd_sfxdevice;
 extern int snd_musicdevice;
 extern int snd_samplerate;
-extern int snd_cachesize;
+extern size_t snd_cachesize;
 extern int snd_maxslicetime_ms;
 extern char* snd_musiccmd;
 extern int snd_pitchshift;
